Fixed issues with optimizations.
I managed to run the optimization tests and the uncovered a narrow use
case where a global variable could be used in both the function argument
and the function body.  I rewrote pieces of the optimizer to more
narrowly detect this and now all tests pass successfully.

Some small improvements for optimizing.

In addition I made all java code get generated with Posix line endings
so it would stop showing differences when rebuilt on Windows.

Recompiled the win32 package and rebuilt the plugin.